Delete Router.jsx
